+++ src/server/sv_types.h
@@ -186,7 +186,7 @@ typedef struct {
 
 	uint32_t frame_latency[SV_CLIENT_LATENCY_COUNT]; // used to calculate ping
 
-	uint32_t frame_size[QUETOO_TICK_RATE]; // used to rate drop packets
+	size_t frame_size[QUETOO_TICK_RATE]; // used to rate drop packets
 	uint32_t rate;
 	uint32_t surpress_count; // number of messages rate suppressed
